requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.24
pydantic>=2.6
python-dotenv>=1.0
orjson>=3.9
typer>=0.12
tqdm>=4.66
lxml>=5.1
beautifulsoup4>=4.12
diskcache>=5.6
numpy>=1.26

langchain>=0.2.10
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0
langchain-text-splitters>=0.3.0

sentence-transformers>=2.7
langchain-huggingface>=1.0.0

faiss-cpu>=1.8.0
pymilvus>=2.4.0
langchain-milvus>=0.3.3

rank-bm25>=0.2.2

# Optional but recommended for fast CSV streaming
pandas>=2.0
